backend/pyproject.toml
[project]
name = "timesheets-backend"
version = "0.1.0"
description = "Timesheet consolidation backend API"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.1",
    "pdfplumber>=0.11.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[tool.setuptools]
py-modules = []
